Split out variables from style.scss file
