+++ .storybook/preview.js
@@ -1,4 +1,5 @@
 import "./style.css";
+import "../agir/front/components/genericComponents/style.scss";
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
