+++ agir/front/components/allPages/index.js
@@ -1,4 +1,5 @@
 import "./style.css";
+import "../genericComponents/style.scss";
 
 import "./renderTopBar";
 import "./help";
