+++ agir/front/components/genericComponents/Card.js
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import style from "./style.scss";
+import style from "./_variables.scss";
 import styled from "styled-components";
 
 const cardTypes = {
